# Performance increase: Only start downloader if needed