The downloader.ps1 job takes quite awhile to start, and gets run on every loop whether it's needed or not. <sup>(actually, it doesn't get started if it's already running, which means the loops were faster *while* it was downloading, then slow again after it finished)</sup>

This only starts it when miners or their prerequisites are missing.  It will still be started when needed, even if miner files were added while the script is running.

Saves about 20-40 seconds on each loop.

Benchmarking:

|System|Before|After (with miners needed)|After (with all miners downloaded|
|---|---|---|---|
|Fast|22628ms|22587ms|4ms|
|Slow|36070ms|37192ms|5ms|